leds: tlc591xx: SW reset during initialization
Our board uses this chip with no connection between the SoC's reset
output and this chip's reset input. If we don't reset the chip and if
nothing initializes all registers in some other manner, there will be
discrepancies in what the LED class' sysfs entries report and the actual
state of the LEDs.

We are also preconfiguring the chip for HW blinking from the boot loader
as an indicator of the I'm-busy-booting state. Once the kernel is up, we
are relying on the DT configuration to blink our status LED in a
different color, and when userspace is up, we finally pass control there
via the standard sysfs entries. Without a full reset from kernel, the
visual result is very confusing and one might get, e.g, purple flashes
instead of a nice solid green light.

Performing a SW reset is a bit tricky because the address for the SW
reset is different than the actual chip's address. This also means that
there's just one shared SWRST address for all chips which on that
particular I2C bus. We cannot therefore send the SWRST during each
chip's probe because doing that would reset all connected chips,
including those which we initialized before. This patch attempts to
solve this by trying to register the SWRST-only I2C address just once.
When probing for other chips, a failure to bind to this shared address
should not be fatal.

This approach will still break on crazy setups with, e.g., a LTC4316
address translator in between the bus master and the slave. I think that
this is acceptable and it's better than trying to bitmask the SWRST out
of the actual address from the DT, for example.

The SWRST address is special -- the datasheet says that the chip won't
ACK anything else the magic SWRST byte sequence. If that's indeed the
case, it might be possible to reuse this address by another I2C chip.
That's another scenario which this patch breaks -- the I2C "device" at
the magic I2C address will be marked as used, and nobody else can use it
without extra force.

Tested with a single TLC59116.

Signed-off-by: Jan Kundrát <jan.kundrat@cesnet.cz>