tty: max310x: Fail probe when external clock crystal is not stable
On (some) of my boards, it appears that it takes up to three *checks* of
this register's value for the external crystal to settle so that it is
reported as "ready" by the chip. For example, on one of these boards it
always succeeds upon a third try no matter if the individual waits are
for 1ms or for 10ms. The original timeout of 10ms is therefore not ideal
as it was generating false warnings on working HW for me. Let's solve
this by retrying up to 20 times (i.e., 200ms).

With this retrying in place, it is now also possible to fail device
initialization altogether. A stable clock is really required in order to
use this UART, so log an error message and bail out if the chip keeps
saying "nope".

Tested on several MAX14830 PCBs.

Signed-off-by: Jan Kundrát <jan.kundrat@cesnet.cz>
---
Changes v1 -> v2:
- fixed uninitialized memory read, thanks to Serge's comment
- use names for constants for repetition/waiting